crud de detalhes finalizado

+++ app/Http/Controllers/Admin/PlanController.php
@@ -60,6 +60,13 @@ class PlanController extends Controller
         if(!$plan)
             return redirect()-back();
 
+
+        if($plan->details->count() > 0 ) {
+            return redirect()
+                    ->back()
+                    ->with('error', 'Existem detalhes vinculados a esses plano, portanto não pode excluir');
+        }
+        
         $plan->delete();
 
         return redirect()->route('plans.index');
